refactor: divide into specific classes
